[IDLE-000] 알림 동의 로직을 Domain에서 제거
BaseFeature의 싱글톤 객체가 담당하도록 구현
